refact: generalize Sort function
