+++ app/dashboard/src/components/Table/Icons.ts
@@ -1,3 +1,4 @@
+
 import { chakra } from '@chakra-ui/react';
 import {
   CheckIcon,
